spotbugs refactor
